feat(vln): add metadata exports to legal and program pages
- Refactor /terms, /privacy, /refunds, and /referrals pages to separate server-side metadata from client-side content
- Create new *Content.tsx components for each page containing interactive elements
- Add proper Metadata exports with OpenGraph configuration for social sharing
- Improves SEO and enables correct Open Graph image generation for all pages
- Ensures metadata consistency with audit standards (Issue #64 - Phase 1: OG & Trust)

Pages updated:
- /terms - Terms of Service with full metadata
- /privacy - Privacy Policy with GDPR compliance details
- /refunds - Refund Policy with detailed service terms
- /referrals - Referral Program with form and commission structure
